ref:change the architecture
