--- src/main/java/io/ylab/habittracker/admin/AdminService.java
@@ -1,16 +1,0 @@
-package io.ylab.habittracker.admin;
-
-import io.ylab.habittracker.habit.model.Habit;
-import io.ylab.habittracker.user.model.User;
-
-import java.util.List;
-
-public interface AdminService {
-    List<User> getUsers();
-
-    List<Habit> getUserHabits(long userId);
-
-    User deleteUser(String email);
-
-    User blockUser(String email);
-}
