--- src/main/java/io/ylab/habittracker/habit/model/Frequency.java
@@ -1,5 +1,0 @@
-package io.ylab.habittracker.habit.model;
-
-public enum Frequency {
-    EVERY_WEEK, EVERY_DAY
-}
