+++ src/main/java/io/ylab/habittracker/model/habit/Frequency.java
@@ -1,0 +1,10 @@
+package io.ylab.habittracker.model.habit;
+
+public enum Frequency {
+    EVERY_WEEK, EVERY_DAY;
+
+    @Override
+    public String toString() {
+        return name().toUpperCase();
+    }
+}
