+++ src/main/java/io/ylab/habittracker/service/admin/AdminService.java
@@ -1,0 +1,16 @@
+package io.ylab.habittracker.service.admin;
+
+import io.ylab.habittracker.model.habit.Habit;
+import io.ylab.habittracker.model.user.User;
+
+import java.util.List;
+
+public interface AdminService {
+    List<User> getUsers();
+
+    List<Habit> getUserHabits(long userId);
+
+    void deleteUser(String email);
+
+    User blockUser(String email);
+}
