--- src/main/java/io/ylab/habittracker/user/model/Status.java
@@ -1,5 +1,0 @@
-package io.ylab.habittracker.user.model;
-
-public enum Status {
-    BLOCKED, AVAIlABLE
-}
